pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "inquisit-bridge"
version = "0.1.0"
description = "Bridge program allowing LLMs to run Inquisit Scripts"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0",
    "imagehash>=4.3",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
inquisit-runner = "runner.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
